# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: dnsaur API
  version: "1.0"
  description: >
    REST API for the dnsaur DNS/DHCP server. All paths are prefixed
    /api/v1. Authenticate with either the dnsaur_session cookie (set by
    POST /auth/login) or an Authorization: Bearer <token> API token
    (see /tokens). API tokens may have scope "read" (GET/HEAD only) or
    "write" (all methods); the session cookie always has full access.
    A read-scoped token used for a POST/PUT/PATCH/DELETE request is
    rejected with 403 {"error": "read-only token"} — see the 403
    response documented on each mutating operation below. Scope is
    enforced by method, with one exception on the read side: a
    read-scoped token gets TSIG secrets redacted (see the TSIGKey
    schema's secret_redacted field), because a GET that returns a
    signing credential in the clear is not something "read-only"
    should cover. The Authorization scheme name is matched
    case-insensitively per RFC 9110 §11.1.
    There is no separate "upstreams" resource: upstream DNS servers are
    configured via the "upstreams" key in /settings (a comma-separated
    list), alongside "upstream.strategy". Each entry is plain by default
    ("1.1.1.1:53") or, with a scheme, DNS-over-TLS
    ("tls://1.1.1.1:853#cloudflare-dns.com") or DNS-over-HTTPS
    ("https://1.1.1.1/dns-query#cloudflare-dns.com") — see the "upstreams"
    key under PUT /settings for the full grammar.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dnsaur_session
      description: Session cookie set by POST /auth/login. HttpOnly, SameSite=Strict.
    bearerAuth:
      type: http
      scheme: bearer
      description: API token minted via POST /tokens, sent as "Authorization Bearer <token>".
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
  responses:
    ReadOnlyToken:
      description: >
        read-only token used for a write operation, or a cookie-authenticated
        write whose Sec-Fetch-Site is cross-site ({"error": "cross-site
        request"}) -- CSRF defence in depth behind SameSite=Strict. Bearer
        requests are never checked for the latter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /groups/{id}/lists:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Groups, Filters]
      summary: Filter lists assigned to a group
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/List" }
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: no group with this id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [Groups, Filters]
      summary: Replace the set of filter lists assigned to a group
      description: >
        Applied as one transaction: an id that names no list changes
        nothing, so the group keeps the assignments it already had.
        Duplicate ids are a set — the same list twice is one assignment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [list_ids]
              properties:
                list_ids:
                  type: array
                  items: { type: integer, format: int64 }
      responses:
        "204":
          description: No Content
        "400":
          description: >
            bad id, invalid json, or `list_ids names list <id>, which does
            not exist`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: no group with this id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /groups/{id}/rules:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [Groups, Filters]
      summary: Per-group allow/block rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer, format: int64 }
                    group_id: { type: integer, format: int64 }
                    action: { type: string, enum: [allow, block] }
                    pattern: { type: string }
                    is_regex: { type: boolean }
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: no group with this id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Groups, Filters]
      summary: Add a rule to a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, pattern]
              properties:
                action: { type: string, enum: [allow, block] }
                pattern:
                  type: string
                  description: >
                    A regex when is_regex is true (max 512 chars, RE2,
                    matched unanchored), otherwise a domain: example.com,
                    *.example.com, or a single label such as localhost.
                    A literal pattern is stored normalised — lowercased,
                    trailing dot and leading "*." removed, Unicode
                    converted to punycode.
                is_regex: { type: boolean, default: false }
      responses:
        "201":
          description: Created
          headers:
            Location:
              description: URL of the rule this request created
              schema: { type: string, example: /api/v1/filters/rules/7 }
          content:
            application/json:
              schema:
                type: object
                description: >
                  The stored rule, whose pattern is the normalised one when
                  the request sent a literal. Its Location is the URL that
                  deletes it: a rule has no read of its own.
                properties:
                  id: { type: integer, format: int64 }
                  group_id: { type: integer, format: int64 }
                  action: { type: string, enum: [allow, block] }
                  pattern: { type: string }
                  is_regex: { type: boolean }
        "400":
          description: >
            invalid json; bad id; action allow|block and pattern required;
            regex pattern too long (max 512); invalid regex; or "pattern
            must be a domain like example.com, *.example.com or localhost"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: no group with this id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /filters/lists:
    get:
      tags: [Filters]
      summary: List all filter lists (block and allow)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/List" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Filters]
      summary: Add a filter list
      description: >
        Adding a list downloads asynchronously after the response (the same
        as POST /filters/refresh): the list row is created and its id
        returned immediately, and the potentially slow full-network refresh
        runs in the background. Every other rule/list/assignment write only
        recompiles from the copies already on disk, and never downloads.


        A fetch is capped at 64 MiB and refuses any address that is not
        public — loopback, link-local and private ranges included, on the
        original URL and on every redirect. A list refused this way reads
        back with last_status "failed" and the reason in last_error.


        The new list is also assigned to every existing group. A group's
        ruleset compiles only from the lists assigned to it, so a list that
        exists but is assigned nowhere filters nothing while still reporting
        itself enabled with a full entry_count.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, kind]
              properties:
                url: { type: string, description: http(s) URL of the list }
                name:
                  type: string
                  maxLength: 120
                  description: >
                    Optional label. Blank or omitted stores blank, which
                    reads back as a name derived from the URL.
                kind: { type: string, enum: [block, allow] }
      responses:
        "201":
          description: Created
          headers:
            Location:
              description: URL of the list this request created
              schema: { type: string, example: /api/v1/filters/lists/4 }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/List"
        "400":
          description: >
            url must be http(s), kind must be block or allow, or name too
            long (max 120)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "409":
          description: that list URL is already subscribed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /filters/lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    patch:
      tags: [Filters]
      summary: Enable, disable, or rename a filter list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                Either field alone is a valid patch, but at least one is
                required. url and kind are deliberately immutable — changing
                either would silently invalidate the on-disk cache and the
                compiled ruleset — and sending one is rejected rather than
                ignored.
              minProperties: 1
              anyOf:
                - required: [enabled]
                - required: [name]
              properties:
                enabled: { type: boolean }
                name:
                  type: string
                  maxLength: 120
                  description: >
                    Blank resets the list to its URL-derived default name
                    rather than clearing it.
      responses:
        "204":
          description: No Content
        "400":
          description: >
            enabled or name required, or name too long (max 120)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [Filters]
      summary: Delete a filter list
      responses:
        "204":
          description: No Content
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /filters/lists/{id}/refresh:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    post:
      tags: [Filters]
      summary: Download one filter list now
      description: >
        Fetches this one list and recompiles, leaving every other
        subscription alone — neither downloaded nor re-stated. Same fetch
        path, same 64 MiB cap, same refusal of any address that is not
        public, and the same last_status / last_error outcome as the
        periodic pass writes.


        202 matches POST /filters/refresh so one verb reads as one status,
        but this one is not fire-and-forget: the download runs inside the
        request, and the body is the list row as the refresh left it. No
        body is read.


        A disabled list is 409: nothing would compile the copy, so the only
        effect would be to overwrite the row's state with the outcome of a
        download that changes nothing.
      responses:
        "202":
          description: Refreshed; the body is the updated list row
          content:
            application/json:
              schema: { $ref: "#/components/schemas/List" }
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: this list is disabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /filters/rules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    delete:
      tags: [Filters]
      summary: Delete a per-group rule
      responses:
        "204":
          description: No Content
        "400":
          description: bad id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: storage unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /filters/refresh:
    post:
      tags: [Filters]
      summary: Trigger an immediate, asynchronous refresh of all filter lists
      responses:
        "202":
          description: Accepted, refresh started in background
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: refreshing }
        "401":
          description: authentication required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          $ref: "#/components/responses/ReadOnlyToken"
...